requirements.txt
aiogram==3.0.0
environs~=9.0
sqlalchemy-orm~=1.2.10
SQLAlchemy~=2.0.20
aiosqlite~=0.19.0
aiogram-dialog~=2.0.0
SQLAlchemy-Utils~=0.41.1
requests~=2.31.0
beautifulsoup4~=4.12.2
python-magic-bin==0.4.14
urllib3~=2.0.5